Globally refactored project structure, removed unnecessary scripts, small refactoring of project files
