--- zeroshoteval/data/loader.py
@@ -1,1 +1,0 @@
-# Script for dataloader construction
