--- zeroshoteval/evaluation/cluster_measurement.py
@@ -1,9 +1,0 @@
-"""
-"""
-
-# region IMPORTS
-# endregion
-
-
-def cluster_measurement_procedure():
-    pass
