--- zeroshoteval/evaluation/verification.py
@@ -1,9 +1,0 @@
-"""
-"""
-
-# region IMPORTS
-# endregion
-
-
-def verification_procedure():
-    pass
