--- zeroshoteval/feature_extraction/attributes_embedding_extractor.py
@@ -1,18 +1,0 @@
-"""
-"""
-
-# import ...
-# ...
-
-from .base_classes._types import EmbeddingObject  # , AttributeObject
-from .base_classes._types import ExtractorType
-from .base_classes.base import EmbeddingExtractor
-from .base_classes.exceptions import ExtractorTypeError
-
-
-class AttributesEmbeddingExtractor(EmbeddingExtractor):
-    # def __init__(self,)
-    # ...
-
-    def extract_embedding():
-        pass
