+++ zeroshoteval/utils/exceptions.py
