--- zeroshoteval/zeroshotnet/gan_net/rungan.py
@@ -1,6 +1,0 @@
-import clswgan
-
-# создали модель, она по болшей части пустая
-GanModel = clswgan.CLSWGAN()
-
-GanModel.initDefoltModel()
